+++ tests/mocks/BootableServiceProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace tests\mocks;
+
+use League\Container\ServiceProvider\BootableServiceProviderInterface;
+use League\Container\ServiceProvider\ServiceProviderInterface;
+
+interface BootableServiceProvider extends
+    BootableServiceProviderInterface
+{
+}